backend/requirements.txt
fastapi                 # Backend web framework
uvicorn                 # ASGI server to run FastAPI
torch                   # PyTorch (used by YOLOv5)
Pillow                  # Image handling (PIL)
numpy                   # For array and image processing
googletrans==4.0.0-rc1  # Translation (if needed)
fpdf                    # PDF report generation
python-multipart        # File uploads in FastAPI
gtts                    # Google Text-to-Speech (for Hindi voice)
python-dotenv           # To manage environment variables (e.g., GitHub token)
seaborn                 # (Optional) For advanced plots or heatmaps
matplotlib              # (Optional) For visualizations
pandas                  # Reading YOLO results, tabular formatting
